aves_peru_2023 Dataset

The aves_peru_2023 dataset comprises a tibble containing information regarding bird species documented in Peru, sourced from the "List of the Birds of Peru" compiled by M. A. Plenge in 2023. It encompasses various details, including scientific names, English names, Spanish names, order, family, and status of each species.

Usage

aves_peru_2023

Format

A tibble with 1,892 rows and 6 columns:

scientific_name

Scientific name of the bird species.

english_name

English common name of the bird species.

spanish_name

Spanish common name of the bird species.

order

The order to which the bird species belongs.

family

The family to which the bird species belongs.

status

Status of the bird species (e.g., resident, endémic, migratory, etc.).

Details

This dataset is designed to provide users with comprehensive information about the avian species found in Peru, as documented by M. A. Plenge. It is organized for easy access and utilization within the R environment.

References

The dataset is based on the "List of the Birds of Peru" by M. A. Plenge.

See also

Author

Data compilation: M. A. Plenge, Package implementation: Paul Efren Santos Andrade

Examples

# Load the avesperu package
library(avesperu)

# Access the avesperu_tab dataset
data("aves_peru_2023")

# Display the first few rows
head(aves_peru_2023)
#> # A tibble: 6 × 6
#>   scientific_name           english_name        spanish_name order family status
#>   <chr>                     <chr>               <chr>        <chr> <chr>  <chr> 
#> 1 Rhea pennata              Lesser Rhea         Ñandú Petizo Rhei… Rheid… nill  
#> 2 Nothocercus julius        Tawny-breasted Tin… Perdiz de P… Tina… Tinam… nill  
#> 3 Nothocercus bonapartei    Highland Tinamou    Perdiz Mont… Tina… Tinam… nill  
#> 4 Nothocercus nigrocapillus Hooded Tinamou      Perdiz de C… Tina… Tinam… nill  
#> 5 Tinamus tao               Gray Tinamou        Perdiz Gris  Tina… Tinam… nill  
#> 6 Tinamus osgoodi           Black Tinamou       Perdiz Negra Tina… Tinam… nill