aves_peru_2024_v2 Dataset

The aves_peru_2024_v2 dataset provides an updated tibble of bird species recorded in Peru, based on the latest taxonomic revisions by the South American Checklist Committee (SACC).

Usage

aves_peru_2024_v2

Format

A tibble with 1,909 rows and 6 columns:

order_name

Taxonomic order to which the bird species belongs.

family_name

Taxonomic family to which the bird species belongs.

scientific_name

Scientific name of the bird species.

english_name

English common name of the bird species.

spanish_name

Spanish common name of the bird species.

status

Category indicating the species' status, based on the following codes: X - Resident species. E - Endemic species. NB - Migratory species. V - Vagrant species. IN - Introduced species. EX - Extirpated species. H - Hypothetical species.

Details

This version incorporates significant taxonomic changes, new categories, and updates derived from recent publications, photographs, and sound recordings archived in accredited institutions.

Total species: 1,909

Distribution by status:

X: 1,542 species

E: 117 species

NB: 138 species

V: 83 species

IN: 3 species

EX: 0 species

H: 26 species

Species near Peru’s borders with Ecuador, Colombia, Brazil, and Chile are included at the end of the list as potential additions to the Peruvian bird inventory. These updates reflect the SACC's ongoing efforts to evaluate and recognize certain subspecies as full species.

References

Plenge, M. A. (2024). Lista de las aves del Perú. Last updated: December 2, 2024. Unión de Ornitólogos del Perú.

See also

Author

Data compilation: Manuel A. Plenge Package implementation: Paul Efren Santos Andrade

Examples

# Load the dataset
data("aves_peru_2024_v2")
#> Warning: data set ‘aves_peru_2024_v2’ not found

# Display the first few rows
head(aves_peru_2024_v2)
#> Error: object 'aves_peru_2024_v2' not found