aves_peru_2025_v1

The aves_peru_2025_v1 dataset provides an updated tibble of bird species recorded in Peru, based on the latest taxonomic revisions by the South American Checklist Committee (SACC).

Usage

aves_peru_2025_v1

Format

A tibble with 1,911 rows and 6 columns:

order_name

Taxonomic order to which the bird species belongs.

family_name

Taxonomic family to which the bird species belongs.

scientific_name

Scientific name of the bird species.

english_name

English common name of the bird species.

spanish_name

Spanish common name of the bird species.

status

Category indicating the species' status, based on the following codes: X - Resident species. E - Endemic species: A species is considered endemic to Peru until a record outside its borders is published. NB - Non-breeding species: Species that occur regularly in Peru but only during their non-breeding period. V - Vagrant species: Erratic species that occasionally occur in Peru and are not part of the usual avifauna. IN - Introduced species: Species introduced to Peru by humans (or have colonized from introduced populations elsewhere) and have established self-sustaining breeding populations. H - Hypothetical species: Records based only on observations, specimens of dubious origin, unpublished photographs, or recordings kept in private hands. EX - Extinct or extirpated species: Species that have gone extinct or have been extirpated from Peru.

Details

This version incorporates significant taxonomic changes, new categories, and updates derived from recent publications, photographs, and sound recordings archived in accredited institutions. Additionally, a classification criterion has been included, following the SACC guidelines.

Total species: 1,911

Distribution by status:

X: 1,542 species

E: 118 species

NB: 138 species

V: 84 species

IN: 3 species

EX: 0 species

H: 26 species

These updates reflect the SACC's ongoing efforts to evaluate and recognize certain subspecies as full species.

References

Plenge, M. A. (2025). Lista de las aves del Perú. Last updated: February 6, 2025. Unión de Ornitólogos del Perú.

See also

Author

Data compilation: Manuel A. Plenge Package implementation: Paul Efren Santos Andrade

Examples

# Load the dataset
data("aves_peru_2025_v1")

# Display the first few rows
head(aves_peru_2025_v1)
#> # A tibble: 6 × 6
#>   order_name   family_name scientific_name      english_name spanish_name status
#>   <chr>        <chr>       <chr>                <chr>        <chr>        <chr> 
#> 1 Rheiformes   Rheidae     Rhea pennata         Lesser Rhea  Ñandú Petizo Resid…
#> 2 Tinamiformes Tinamidae   Nothocercus julius   Tawny-breas… Perdiz de P… Resid…
#> 3 Tinamiformes Tinamidae   Nothocercus bonapar… Highland Ti… Perdiz Mont… Resid…
#> 4 Tinamiformes Tinamidae   Nothocercus nigroca… Hooded Tina… Perdiz de C… Resid…
#> 5 Tinamiformes Tinamidae   Tinamus tao          Gray Tinamou Perdiz Gris  Resid…
#> 6 Tinamiformes Tinamidae   Tinamus osgoodi      Black Tinam… Perdiz Negra Resid…