avesperu_tab Dataset

The avesperu_tab dataset contains a tibble that provides information on bird species recorded in Peru, based on the "List of the Birds of Peru" by M. A. Plenge. It includes details such as scientific names, English names, Spanish names, order, family, and status of each species.

avesperu_tab

Format

A tibble with 1,892 rows and 6 columns:

scientific_name

Scientific name of the bird species.

english_name

English common name of the bird species.

spanish_name

Spanish common name of the bird species.

order

The order to which the bird species belongs.

family

The family to which the bird species belongs.

status

Status of the bird species (e.g., resident, endémic, migratory, etc.).

Details

This dataset is designed to provide users with comprehensive information about the avian species found in Peru, as documented by M. A. Plenge. It is organized for easy access and utilization within the R environment.

References

The dataset is based on the "List of the Birds of Peru" by M. A. Plenge.

See also

Author

Data compilation: M. A. Plenge, Package implementation: Paul Efren Santos Andrade

Examples

# Load the avesperu package
library(avesperu)

# Access the avesperu_tab dataset
data("avesperu_tab")

# Display the first few rows
head(avesperu_tab)
#>             scientific_name           english_name            spanish_name
#> 1              Rhea pennata            Lesser Rhea            Ñandú Petizo
#> 2        Nothocercus julius Tawny-breasted Tinamou Perdiz de Pecho Leonado
#> 3    Nothocercus bonapartei       Highland Tinamou          Perdiz Montesa
#> 4 Nothocercus nigrocapillus         Hooded Tinamou  Perdiz de Cabeza Negra
#> 5               Tinamus tao           Gray Tinamou             Perdiz Gris
#> 6           Tinamus osgoodi          Black Tinamou            Perdiz Negra
#>          order    family status
#> 1   Rheiformes   Rheidae   nill
#> 2 Tinamiformes Tinamidae   nill
#> 3 Tinamiformes Tinamidae   nill
#> 4 Tinamiformes Tinamidae   nill
#> 5 Tinamiformes Tinamidae   nill
#> 6 Tinamiformes Tinamidae   nill